initial support for ntfs partition

+++ create_image.sh
@@ -84,7 +84,7 @@ echo "-> rootfs size: ${ROOTFS_SIZE}kb"
 
 echo "Creating empty image in $DEST_IMAGE"
 #dd if=/dev/zero of="$DEST_IMAGE" bs=1M count=1024 conv=sync,fsync >/dev/null 2>&1
-fallocate -l 2G "$DEST_IMAGE" >/dev/null 2>&1
+fallocate -l 512M "$DEST_IMAGE" >/dev/null 2>&1
 
 if [ $? -ne 0 ]; then
 	echo "Error while creating $DEST_IMAGE empty file"
@@ -106,22 +106,31 @@ if [ $? -ne 0 ]; then
 	exit 3
 fi
 
-START=$BEGIN_USER_PARTITIONS
-END=$(($START + $ROOTFS_SECTORS - 1))
-parted -s -- "$LOOP_DEVICE" unit s mkpart primary fat32 $(($END + 1)) -1s >/dev/null 2>&1
+START_ROOTFS=$BEGIN_USER_PARTITIONS
+END_ROOTFS=$(($START_ROOTFS + $ROOTFS_SECTORS - 1))
+START_FAT=$(($END_ROOTFS + 1))
+END_FAT=$(($START_FAT + 131072)) # 131072 sectors = 64Mb
+START_NTFS=$(($END_FAT + 1))
+parted -s -- "$LOOP_DEVICE" unit s mkpart primary fat32 $START_FAT $END_FAT >/dev/null 2>&1
 if [ $? -ne 0 ]; then
 	echo "Could not create fat partition"
 	exit 3
 fi
 
-parted -s -- "$LOOP_DEVICE" unit s mkpart primary $START $END >/dev/null 2>&1 
+parted -s -- "$LOOP_DEVICE" unit s mkpart primary ntfs $START_NTFS -1s >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+	echo "Could not create ntfs partition"
+	exit 3
+fi
+
+parted -s -- "$LOOP_DEVICE" unit s mkpart primary $START_ROOTFS $END_ROOTFS >/dev/null 2>&1
 if [ $? -ne 0 ]; then
 	echo "Could not create rootfs partition"
 	exit 3
 fi
 
 
-parted -s -- "$LOOP_DEVICE" set 1 boot on set 2 boot off >/dev/null 2>&1
+parted -s -- "$LOOP_DEVICE" set 1 boot on set 2 boot off set 3 boot off >/dev/null 2>&1
 if [ $? -ne 0 ]; then
 	echo "Could not set partition flags"
 	exit 28
@@ -133,12 +142,14 @@ sleep 1
 # First check: in containers, it may happen that loop device partitions
 # spawns as soon as they are created. We check their presence. If they already
 # are there, we don't remount the device
-SQUASHFS_PARTITION="${LOOP_DEVICE}p2"
+SQUASHFS_PARTITION="${LOOP_DEVICE}p3"
+NTFS_PARTITION="${LOOP_DEVICE}p2"
 FAT_PARTITION="${LOOP_DEVICE}p1"
 echo "squashfs partition: $SQUASHFS_PARTITION"
 echo "fat partition: $FAT_PARTITION"
+echo "ntfs partition: $NTFS_PARTITION"
 
-if [ ! -b "$SQUASHFS_PARTITION" -o ! -b "$FAT_PARTITION" ]; then
+if [ ! -b "$SQUASHFS_PARTITION" -o ! -b "$FAT_PARTITION" -o ! -b "$NTFS_PARTITION" ]; then
 	echo "Remounting loop device with partitions"
 	losetup -d "$LOOP_DEVICE" >/dev/null 2>&1
 	sleep 1
@@ -154,10 +165,12 @@ if [ ! -b "$SQUASHFS_PARTITION" -o ! -b "$FAT_PARTITION" ]; then
 		exit 5
 	fi
 
-	SQUASHFS_PARTITION="${LOOP_DEVICE}p2"
+	SQUASHFS_PARTITION="${LOOP_DEVICE}p3"
+	NTFS_PARTITION="${LOOP_DEVICE}p2"
 	FAT_PARTITION="${LOOP_DEVICE}p1"
 	echo "squashfs partition after remount: $SQUASHFS_PARTITION"
 	echo "fat partition: after remount $FAT_PARTITION"
+	echo "ntfs partition: after remount $NTFS_PARTITION"
 
 	sleep 1
 fi
@@ -173,6 +186,11 @@ if [ ! -b "$FAT_PARTITION" ]; then
 	exit 6
 fi
 
+if [ ! -b "$NTFS_PARTITION" ]; then
+	echo "Could not find expected partition $NTFS_PARTITION"
+	exit 6
+fi
+
 echo "Copying squashfs rootfilesystem"
 dd if="${DIST_PATH}/root.img" of="$SQUASHFS_PARTITION" bs=256k conv=sync,fsync >/dev/null 2>&1
 
@@ -185,16 +203,87 @@ fi
 source "${TS_SOURCES_PATH}/boot_install"
 
 echo "Formatting FAT32 partition"
-mkfs.vfat -s 32 -n "MULTITOOL" "$FAT_PARTITION" >/dev/null 2>&1
+mkfs.vfat -s 16 -n "BOOTSTRAP" "$FAT_PARTITION" >/dev/null 2>&1
 
 if [ $? -ne 0 ]; then
-	echo "Could not format partition"
+	echo "Could not format FAT32 partition"
 	exit 7
 fi
 
-echo "Mounting FAT32 partition"
+echo "Formatting NTFS partition"
+mkfs.ntfs -f -L "MULTITOOL" -s 4096 -c 16384 "$NTFS_PARTITION" >/dev/null 2>&1
+
+if [ $? -ne 0 ]; then
+	echo "Could not format NTFS partition"
+	exit 7
+fi
+
 TEMP_DIR=$(mktemp -d)
 
+echo "Mounting NTFS partition"
+mount "$NTFS_PARTITION" "$TEMP_DIR" >/dev/null 2>&1
+
+if [ $? -ne 0 ]; then
+	echo "Could not mount $NTFS_PARTITION to $TEMP_DIR"
+	exit 9
+fi
+
+echo "Populating partition"
+
+cp "${CWD}/LICENSE" "${TEMP_DIR}/LICENSE"
+if [ $? -ne 0 ]; then
+	echo "Could not copy LICENSE to partition"
+	exit 28
+fi
+
+git log --no-merges --pretty="%as: %s" > "${TEMP_DIR}/CHANGELOG"
+if [ $? -ne 0 ]; then
+	echo "Could not store CHANGELOG to partition"
+fi
+
+git log -1 --pretty="%h - %aD" > "${TEMP_DIR}/ISSUE"
+if [ $? -ne 0 ]; then
+	echo "Could not store ISSUE to paritition"
+fi
+
+echo "${TARGET_CONF}" > "${TEMP_DIR}/TARGET"
+if [ $? -ne 0 ]; then
+	echo "Could not store TARGET to partition"
+fi
+
+mkdir -p "${TEMP_DIR}/backups"
+if [ $? -ne 0 ]; then
+	echo "Could not create backup directory"
+	exit 28
+fi
+
+mkdir -p "${TEMP_DIR}/images"
+if [ $? -ne 0 ]; then
+	echo "Could not create images directory"
+	exit 29
+fi
+
+mkdir -p "${TEMP_DIR}/bsp"
+if [ $? -ne 0 ]; then
+	echo "Could not create bsp directory"
+	exit 30
+fi
+
+echo "Copying board support package blobs into bsp directory"
+cp "${DIST_PATH}/uboot.img" "${TEMP_DIR}/bsp/uboot.img"
+
+[[ -f "${DIST_PATH}/trustos.img" ]] && cp "${DIST_PATH}/trustos.img" "${TEMP_DIR}/bsp/trustos.img"
+[[ -f "${DIST_PATH}/legacy-uboot.img" ]] && cp "${DIST_PATH}/legacy-uboot.img" "${TEMP_DIR}/bsp/legacy-uboot.img"
+
+echo "Unmount NTFS partition"
+umount "$NTFS_PARTITION" >/dev/null 2>&1
+
+if [ $? -ne 0 ]; then
+	echo "Could not umount $NTFS_PARTITION"
+	exit 17
+fi
+
+echo "Mounting FAT32 partition"
 if [ $? -ne 0 ]; then
 	echo "Could not create temporary directory"
 	exit 8
@@ -232,30 +321,6 @@ if [ $? -ne 0 ]; then
 	exit 14
 fi
 
-mkdir -p "${TEMP_DIR}/backups"
-if [ $? -ne 0 ]; then
-	echo "Could not create backup directory"
-	exit 28
-fi
-
-mkdir -p "${TEMP_DIR}/images"
-if [ $? -ne 0 ]; then
-	echo "Could not create images directory"
-	exit 29
-fi
-
-mkdir -p "${TEMP_DIR}/bsp"
-if [ $? -ne 0 ]; then
-	echo "Could not create bsp directory"
-	exit 30
-fi
-
-echo "Copying board support package blobs into bsp directory"
-cp "${DIST_PATH}/uboot.img" "${TEMP_DIR}/bsp/uboot.img"
-
-[[ -f "${DIST_PATH}/trustos.img" ]] && cp "${DIST_PATH}/trustos.img" "${TEMP_DIR}/bsp/trustos.img"
-[[ -f "${DIST_PATH}/legacy-uboot.img" ]] && cp "${DIST_PATH}/legacy-uboot.img" "${TEMP_DIR}/bsp/legacy-uboot.img"
-
 # Gather the PARTUUID of the squashfs partition loop device
 # blkid is friendlier in case of containers, so we use it here in place of lsblk
 SQUASHFS_PARTITION_UUID=$(blkid -o value -s PARTUUID $SQUASHFS_PARTITION)
@@ -292,27 +357,6 @@ if [ $? -ne 0 ]; then
 	exit 16
 fi
 
-cp "${CWD}/LICENSE" "${TEMP_DIR}/LICENSE"
-if [ $? -ne 0 ]; then
-	echo "Could not copy LICENSE to partition"
-	exit 28
-fi
-
-git log --no-merges --pretty="%as: %s" > "${TEMP_DIR}/CHANGELOG"
-if [ $? -ne 0 ]; then
-	echo "Could not store CHANGELOG to partition"
-fi
-
-git log -1 --pretty="%h - %aD" > "${TEMP_DIR}/ISSUE"
-if [ $? -ne 0 ]; then
-	echo "Could not store ISSUE to paritition"
-fi
-
-echo "${TARGET_CONF}" > "${TEMP_DIR}/TARGET"
-if [ $? -ne 0 ]; then
-	echo "Could not store TARGET to partition"
-fi
-
 echo "Unmount FAT32 partition"
 umount "$FAT_PARTITION" >/dev/null 2>&1
 
@@ -336,7 +380,7 @@ if [ $? -ne 0 ]; then
 	exit 23
 fi
 
-truncate -s 140M "$DEST_IMAGE"
+#truncate -s 140M "$DEST_IMAGE"
 
 sync
